audit: pin sevtChngCamera/sevtPause GG variable sizing for LP64 safety
Same bug class as kcbVarFreeze: GG variable parts sized via
sizeof(long) / sizeof(long)*2 in scene.cpp shift on LP64
(long==8). The on-disk slots are 4 bytes (camera index) and 8
bytes (SceneEventPause = WaitReason + int32_t).

Fixes:
  - sevtChngCamera variable: 4 sites changed from size(long) to
    size(int32_t) -- _FAddSev write, the two pggFrm/StartTemp
    save-path FInsert sites, and the dumper read in movie_chomp.
  - sevtPause _FAddSev: size(long) * 2 -> size(SceneEventPause).
  - GG read/write locals (icam, icamNext, picamOld) routed
    through int32_t scratch where the public function signature
    keeps `long` for source-compat. The runtime GG entries are
    now 4 bytes everywhere; the scratch widens to long at the
    call boundary.
  - movie_chomp dumper reads sevtChngCamera variable as int32_t
    (was long), %d format.

Wire format unchanged on x86; LP64 now consistent with x86 byte
count.

+++ src/tools/movie_chomp.cpp
@@ -1186,11 +1186,10 @@ bool MovieDecompiler::FDecompile(PChunkyFile pcflSrc, PMSNK pmsnk, PMSNK pmsnkEr
                     }
 
                     case sevtChngCamera: {
-                        long iangle;
-                        iangle = *(long*)pggsevStart->QvGet(isevStart);
-                        // long
+                        int32_t iangle;
+                        iangle = *(int32_t*)pggsevStart->QvGet(isevStart);
                         printf("\t\tsevt=sevtChngCamera,\n");
-                        printf("\t\tiangle=%lu\n", iangle);
+                        printf("\t\tiangle=%d\n", iangle);
                         break;
                     }
